Hg-commit: 434c31c7a601b1c2168f3762239bf932db893dbc

+++ include/sieve/ALE_mem.hh
@@ -25,8 +25,8 @@ namespace ALE {
   class MemoryLogger {
   public:
     struct Log {
-      int num;
-      int total;
+      long long num;
+      long long total;
       std::map<std::string, int> items;
 
       Log(): num(0), total(0) {};
